+++ Pandora/Paradigm/Structure/Some/List.hs
@@ -39,7 +39,7 @@ import Pandora.Paradigm.Structure.Ability.Focusable (Focusable (Focusing, focusi
 import Pandora.Paradigm.Structure.Ability.Deletable (Deletable ((-=)))
 import Pandora.Paradigm.Structure.Ability.Measurable (Measurable (Measural, measurement), Scale (Length), measure)
 import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic (reduce, resolve))
-import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morph (Rotate, Into, Prepend), premorph, rotate, prepend)
+import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morph (Rotate, Into, Insert), premorph, rotate, insert)
 import Pandora.Paradigm.Structure.Ability.Substructure (Substructure (Substructural, substructure), Segment (Tail), sub, subview)
 import Pandora.Paradigm.Structure.Interface.Stack (Stack (push, pop))
 
@@ -61,14 +61,14 @@ instance Stack List where
 	push x = lift . Construct x . run
 	pop xs = resolve deconstruct Nothing ||= xs
 
-instance Morphable (Prepend Left) List where
-	type Morphing (Prepend Left) List = T_U Covariant Covariant (->) Identity List
+instance Morphable (Insert Left) List where
+	type Morphing (Insert Left) List = T_U Covariant Covariant (->) Identity List
 	morphing (premorph -> xs) = T_U $ \(Identity x) -> lift . Construct x . run $ xs
 
 instance Focusable Head List where
 	type Focusing Head List a = Maybe a
 	focusing (extract -> stack) = Store $ extract <$> run stack :*: \case
-		Just x -> stack & subview @Tail & prepend @Left x & Tag
+		Just x -> stack & subview @Tail & insert @Left x & Tag
 		Nothing -> stack & subview @Tail & Tag
 
 instance Measurable Length List where
@@ -112,8 +112,8 @@ instance Focusable Head (Construction Maybe) where
 	type Focusing Head (Construction Maybe) a = a
 	focusing (extract -> stack) = Store $ extract stack :*: Tag . Construct % deconstruct stack
 
-instance Morphable (Prepend Left) (Construction Maybe) where
-	type Morphing (Prepend Left) (Construction Maybe) = T_U Covariant Covariant (->) Identity (Construction Maybe)
+instance Morphable (Insert Left) (Construction Maybe) where
+	type Morphing (Insert Left) (Construction Maybe) = T_U Covariant Covariant (->) Identity (Construction Maybe)
 	morphing (premorph -> xs) = T_U $ \(Identity x) -> Construct x $ Just xs
 
 instance Measurable Length (Construction Maybe) where
@@ -138,24 +138,24 @@ instance {-# OVERLAPS #-} Extendable (Tap ((:*:) <:.:> List)) where
 instance Morphable (Rotate Left) (Tap ((:*:) <:.:> List)) where
 	type Morphing (Rotate Left) (Tap ((:*:) <:.:> List)) = Maybe <:.> Zipper List
 	morphing (premorph -> Tap x (T_U (bs :*: fs))) = TU
-		$ Tap % twosome (subview @Tail bs) (prepend @Left x fs) <$> view (focus @Head) bs
+		$ Tap % twosome (subview @Tail bs) (insert @Left x fs) <$> view (focus @Head) bs
 
 instance Morphable (Rotate Right) (Tap ((:*:) <:.:> List)) where
 	type Morphing (Rotate Right) (Tap ((:*:) <:.:> List)) = Maybe <:.> Zipper List
 	morphing (premorph -> Tap x (T_U (bs :*: fs))) = TU
-		$ Tap % twosome (prepend @Left x bs) (subview @Tail fs) <$> view (focus @Head) fs
+		$ Tap % twosome (insert @Left x bs) (subview @Tail fs) <$> view (focus @Head) fs
 
 type instance Zipper (Construction Maybe) = Tap ((:*:) <:.:> Construction Maybe)
 
 instance Morphable (Rotate Left) (Tap ((:*:) <:.:> Construction Maybe)) where
 	type Morphing (Rotate Left) (Tap ((:*:) <:.:> Construction Maybe)) = Maybe <:.> Zipper (Construction Maybe)
 	morphing (premorph -> Tap x (T_U (bs :*: fs))) = TU
-		$ Tap (extract bs) . twosome % (prepend @Left x fs) <$> deconstruct bs
+		$ Tap (extract bs) . twosome % (insert @Left x fs) <$> deconstruct bs
 
 instance Morphable (Rotate Right) (Tap ((:*:) <:.:> Construction Maybe)) where
 	type Morphing (Rotate Right) (Tap ((:*:) <:.:> Construction Maybe)) = Maybe <:.> Zipper (Construction Maybe)
 	morphing (premorph -> Tap x (T_U (bs :*: fs))) = TU
-		$ Tap (extract fs) . twosome (prepend @Left x bs) <$> deconstruct fs
+		$ Tap (extract fs) . twosome (insert @Left x bs) <$> deconstruct fs
 
 instance Monotonic a (Maybe <:.> Construction Maybe := a) where
 	reduce f r = reduce f r . run